update toolchain to 2023-04-15, the same as the most recent release of rust-gpu. Also updated crate dependencies to be compatible with this release.
Using this toolchain means we're stuck to an older version of bindgen. In newer versions of llvm, characters for identifiers of files are emitted that bindgen says are invalid. This was fixed in bindgen 0.62, but since we're on 0.58, we don't have that bug fix and need to use llvm version 15 or older