refactor(cust_raw): Add `must_use` attrbuites on result/error types.
CUDA libraries have different result types (`CUresult`, `cudaError_t`, etc.) that must be checked after calling the function. This change adds `must_use` attribute on those types and functions that return that type when generating the bindings. This also fixes the warnings caused by this in places where the result type was not checked by either checking the result type, or ignore the value.