Spill large statics from constant to global memory
This isn't fully correct, as ideally we keep track
of what we have put into constant memory and when
it is filled up spill instdead of only spilling when
a static is big. But, this is materially better than
what is there (which is a runtime error).

An argument can be made to just _always_ use global
memory and we don't have to worry about getting the
packing right.

Fixes https://github.com/Rust-GPU/Rust-CUDA/issues/208.

See also the debugging and discussion in
https://github.com/Rust-GPU/Rust-CUDA/pull/216