Add compiletests
These tests check that code compiles (or doesn't) as expected.
The framework also has the capability to check the generated ptx (see
the tests in `dis/`.

These tests are not run, so they do not validate runtime behavior.
This also means they can run on machines without a GPU, like GitHub
Actions runners.

This is largely modeled / structured after rust-gpu's compiletests
in anticipation of a project merge one day.