Improve and rename `inbounds!` macro.
It has two versions, one with an upper bound, and one with a lower and
upper bound. This commit removes the first one and changes the second
one to take a range, because that is more concise and flexible and
clearer.

Also, rename it as `in_range`, which makes sense given that the bounds
are specified via a Rust `Range`.

Note: some of the ranges are incorrect, and will be fixed in the next
commit.