Increase the minimum CUDA version to 12.0 and the default arch to `compute_75`.
CUDA 12.0 was released in December 2022, and CUDA 13.0 was released in
August 2025. It feels like a good time to drop CUDA 11.x support. This
means later Kepler devices with compute capabilities of 3.5 and 3.7 will
no longer be supported. The minimum version of NVVM IR increases from
1.6 to 2.0, because that's what CUDA 12.0 uses.

Along with this, change the default compiler target to `compute_75`.
This is a good choice because it's the minimum supported by CUDA 13.0,
and gets Rust CUDA a step closer to working with CUDA 13.0.

The existing defaults were all over the place.
- `NvvmArch::default()` was `compute_52`.
- `CudaBuilder`'s default was `compute_61`.
- compiletest's default was `compute_70`.

This commit makes the latter two determined by `NvvmArch::default()`,
which is changed to `compute_75`.

Currently CI runs compiletests on `compute_61`, `compute_70`, and
`compute_90`; this commit changes the `compute_70` to `compute_75`. It
seems sensible to have the default value as one of the things tested by
CI.

This comment also adds a comment on NvvmArch with a table of
CUDA/`compute_*` values, which I found very useful.

Resources:
- https://en.wikipedia.org/wiki/CUDA#GPUs_supported for compute
  capabilities supported by different CUDA versions.
- https://docs.nvidia.com/cuda/archive/12.0.0/cuda-toolkit-release-notes/index.html
  for NVVM IR version information.