List workspace members explicitly.
Currently the top-level `Cargo.toml` lists members both explicitly
(direct naming) and implicitly (using globs). The latter is annoying,
and makes it hard to tell what the actual packages in the workspace are,
particularly with the complex `cargo --workspace --exclude` invocations
in CI.

This commit makes all the member explicit and puts them in alphabetical
order.