Remove `JitTarget`.
It includes some now-unsupported targets and is also missing some new
targets. The obvious thing to do is update it, but it's simpler and
better to recognize that it's identical to the generated
`driver_sys::CUjit_target` type and instead use that generated type
directly, avoiding the need for manual updating in the future.
Especially given that there is a non-trivial encoding for targets with
'a' and 'f' suffixes (which involves adding 2^16 and 2^17, respectively,
to the base number).

This seems fine because this `ModuleJitOption` type is obscure and has
no existing uses in the codebase.