Re-enable `path-tracer`.
In #313 I disabled all the code relying on OptiX. I subsequently learned
that `path-tracer`'s use of OptiX is optional. `path-tracer` is a big
and good example, so this commit re-enables it, but with the
OptiX-specific bits commented out. (Commenting out code is crude, but I
had trouble with features, as detailed within the commit. This is better
than nothing.)